feat: :sparkles: Hook made for API.
Custom API hook. Env created to store api info. Constant folder created.
