+++ .env
@@ -1,0 +1,1 @@
+VITE_AMBITO_URL=https://mercados.ambito.com/home/general3
