+++ src/components/Navbar.tsx
@@ -1,35 +1,15 @@
-import { useQuery } from 'react-query';
 import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import { grey } from '@mui/material/colors';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
 
+import useFetchDolarValue from '~hooks/useFetchDolarValue';
+
 import PriceDisplay from './PriceDisplay';
 
 export default function Navbar() {
-  const DOLAR_REFETCH_INTERVAL = 300000;
-  const fetchDolarValue = async () => {
-    const response = await fetch('https://mercados.ambito.com/home/general3');
-    return response.json();
-  };
-
-  const { data, status } = useQuery({
-    queryKey: ['dolarValue'],
-    queryFn: fetchDolarValue,
-    refetchInterval: DOLAR_REFETCH_INTERVAL,
-  });
-
-  if (status === 'loading') {
-    return <div>Loading...</div>;
-  }
-
-  if (status === 'error') {
-    return <div>Error: Failed to fetch Dolar value.</div>;
-  }
-
-  const DOLAR_INFORMAL = data[1].venta;
-
+  const DOLAR_INFORMAL = useFetchDolarValue();
   return (
     <Box sx={{ mt: 1.5 }}>
       <AppBar position='static' elevation={0}>
